Fix invalid Fragment state restoring
When FragmentStatePagerAdapter restores fragment because of setPrimaryItem
call, the restored fragment's children are instantiated before their parent
is added and marked as active. As a result, the mWho parameter in child
Fragments is invalid. This parameter is used to get the LoaderManager for
the Fragment. This leads to situation when two Fragments operate on the
same LoaderManager and ids conflict is possible.

I'm not sure if I have fixed it in a correct place. Whole Loader/Fragment
implementation is a huge clusterfuck I do not pretend to understand fully,
so I did the simplest thing I could do.