Fix issue with delivering Loader results
In case the loader is restarted twice after the original loader have
successfully loaded and delivered the data and the background operation
for the 1st restart has finished, the 2nd restart was never executed.

The 2nd restart was added to current LoaderInfo as mPendingLoader, but
it was never installed in LoaderManager, because the `info.cancel()`
was called before the mPendingLoader was set. It worked if the background
operation took longer, because then the 2nd restart was executed from
AsyncTaskLoader.dispatchOnCancelled().

I'm not sure it's a correct fix. Another option would be changing the
`AsyncTask.dispatchOnCancelled()` method.